(PC-20534) home/components: move computeDateRangeDisplay helper
