+++ src/features/home/components/modules/ThematicHighlightModule.tsx
@@ -4,7 +4,7 @@ import React, { FunctionComponent, useEffect } from 'react'
 import styled from 'styled-components/native'
 
 import { HIGHLIGHT_TEXT_BACKGROUND_OPACITY } from 'features/home/components/constants'
-import { computeDateRangeDisplay } from 'features/home/components/modules/helpers/computeDateRangeDisplay'
+import { computeDateRangeDisplay } from 'features/home/components/helpers/computeDateRangeDisplay'
 import { ThematicHighlightGradient } from 'features/home/components/ThematicHighlightGradient'
 import { getNavigateToThematicHomeConfig } from 'features/navigation/helpers/getNavigateToThematicHomeConfig'
 import { ContentTypes } from 'libs/contentful'
